// Demo/src/types/enums/custom-property.enum.ts
/**
 * Custom properties created in the resource editor are listed here
 */
export declare const enum CustomProperty {
    IsHelpText = 'IsHelpText',
	
    SelectFromMap = 'SelectFromMap',
    Latitude = 'Latitude',
    Longitude = 'Longitude',
    Zoom = 'Zoom',
    ShowUserLocation = 'ShowUserLocation',
}